Add support for Caller Name Display
Copied relevant parts from CdmaConnection.java to get the information
as/if presented by network operator (Fido/Rogers in canada provide that
information, and it get displayed if there's no match in phonebook).
Although it works (the nale is displayed when the call is received),
the information is not stored (not kept in the log).

* Patch Set 2 : correct coding style
* Patch Set 3 : promote [get]cnapName, [get]cnapNapePresentation to parent
                class (Connection)
* Patch Set 4 : -remove ref to CDMA in get[CnapName|CnapNamePresentation] headers
                -remove unnecessary self ref-reference
* Patch Set 5 : -add forgoten "protected" modifier to newly promoted class memebers

Change-Id: I2279f75d679d5afa716d9904fc3b3e33c77948c6