Adding disconnect call to the SurfaceTextureTarget
Related to b/5873421.
When an app using the SurfaceTextureTarget filter and the activity goes
out of focus such as onPause(), the underlying surfacetexture might not
be available anymore to the filter. The filter needs to be told about
that so that it does try to render into the surface corresponding to the
surfacetexture. For example, in Camera app, the VideoCamera activity has
to let the underlying effects graph know about the disconnect in the
onPause() method.

Change-Id: Iee3af16715432d84fef4438d1671f4d6d261b3d7