Change SMS security check to enforceCallingPermission().
The permission check for IccSmsInterfaceManager.sendText() was changed from
enforceCallingPermission() to enforceCallingOrSelfPermission() to enable the
Phone app to send SMS's directly for the "reply by SMS" feature. That feature
was implemented in a different way (handled by the MMS app), so the permission
check can be changed back to the original enforceCallingPermission(). Verified
that SMS can still be sent, including "reply by SMS" for an incoming call.

Bug: 4686733
Change-Id: If844a0c485de0a87857d8f82a3452e776005153e