Better support for HTML5 audio loop.
Loop is trigger by a seek to 0 when ended on native side but there is no play
call. So on java side, we detect this and call into native side to trigger a
play after completion.
This fixed the UI problem and keep in sync with the native mode.
Beyond that, we don't need to reload for looping and we don't have the seek
to play artifacts.

bug:5461143

webkit change:
https://android-git.corp.google.com/g/#/c/193750/

Change-Id: I779f3e1fbc789832a1a99d1f17823db6b57b35df