UI changes in suggestion menu
Icons removed for delete and add to dictionary.
Words capitalized (in English only)
Changed background color to white for words.

Note that due to a bug in the interaction between listView and
popup window, setting a background to the TextViews paints a
solid rectangle, which hides the rounded corners defined in
the popup window background.

Bug 6712892

Change-Id: Id5bb429463e2debc5e30d571c34426dfa82119d4