Maybe fix issue #6756472: Share button is not defocusing the shade
I got this reproduce a few times, then wasn't able to.  I made this
change and then couldn't reproduce it.  Maybe it fixed it.

The change here is to explicitly pass in the handler class to
apply() and reapply(), instead of relying it being set as a member of
the RemoteViews class.  This is much cleaner and more explicitly about
setting that for the click callbacks.

Bug: 6756472
Change-Id: I8d029c3836501df3206c433838124b4be3890a8b